Init header

+++ src/algorithms/sort/insertion sort/include/insertion_sort.h
@@ -1,3 +1,5 @@
 #pragma once
 
-void insertion_sort(int ary[], const unsigned int n);
+#include <stdlib.h>
+
+void insertion_sort(size_t ary[], const size_t n);
